+++ packages/mochi-rest/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @consolelabs/mochi-rest
 
+## 3.0.9
+
+### Patch Changes
+
+- UI tweak
+
+## 3.0.8
+
+### Patch Changes
+
+- Remove log
+
 ## 3.0.7
 
 ### Patch Changes